В процессе выполнения Таск 7
